added MaxTempMapper

+++ ch02/src/main/java/com/xieyuanpeng/hadoop/book/ch02/MaxTempMapper.java
@@ -1,5 +1,6 @@
 package com.xieyuanpeng.hadoop.book.ch02;
 
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
@@ -10,7 +11,7 @@ import java.io.IOException;
  * @author xyp
  * @date 2018/11/21
  **/
-public class MaxTempMapper extends Mapper<LongWritable, Text, Text, LongWritable> {
+public class MaxTempMapper extends Mapper<LongWritable, Text, Text, IntWritable> {
     public static final int MISSING = 9999;
 
     @Override
@@ -19,7 +20,7 @@ public class MaxTempMapper extends Mapper<LongWritable, Text, Text, LongWritable
         String year = lineWords[0];
         int airTemp = Integer.parseInt(lineWords[4]);
         if (airTemp != MISSING) {
-            context.write(new Text(year), new LongWritable(airTemp));
+            context.write(new Text(year), new IntWritable(airTemp));
         }
     }
 }
